+++ src/components/MapLabel.svelte
@@ -2,11 +2,20 @@
   import { faXmark } from '@fortawesome/free-solid-svg-icons';
   import Fa from 'svelte-fa/src/fa.svelte';
   import MapStyleInputWrapper from './MapStyleInputWrapper.svelte';
-  import { showDisplays as showDisplaysStore } from '../stores';
+  import MapLocationControl from './MapLocationControl.svelte';
+  import {
+    showDisplays as showDisplaysStore,
+    linkLocations as linkLocationsStore,
+  } from '../stores';
   export let index;
   export let name;
   export let onClose;
   export let disableClose;
+  export let mapState;
+
+  let locationProps;
+
+  $: ({ showBoundaries, showCollisions, ...locationProps } = mapState);
 </script>
 
 {#if $showDisplaysStore}
@@ -15,7 +24,14 @@
       <Fa icon={faXmark} />
     </button>
     <div class="map-name">{name}</div>
-    <MapStyleInputWrapper {index} />
+    <div class="options-container">
+      <MapStyleInputWrapper {index} />
+      {#if !$linkLocationsStore && Object.keys(locationProps).length}
+        <div class="location-control">
+          <MapLocationControl on:mapState {...locationProps} />
+        </div>
+      {/if}
+    </div>
   </div>
 {/if}
 
@@ -51,5 +67,18 @@
   .close-button:disabled {
     cursor: not-allowed;
     color: lightgray;
+  }
+
+  .options-container {
+    display: flex;
+    flex-direction: column;
+    align-items: left;
+    justify-content: center;
+  }
+
+  .location-control {
+    margin-top: 12px;
+    display: flex;
+    justify-content: left;
   }
 </style>