+++ src/stores.js
@@ -6,3 +6,7 @@ export const stylePresets = writable([]);
 export const config = writable([]);
 
 export const showDisplays = writable(true);
+
+export const linkLocations = writable(true);
+
+export const mapLocations = writable([]);